requirements.txt
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.0
dnacentersdk==2.6.11
et-xmlfile==1.1.0
fastjsonschema==2.19.1
future==0.18.3
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
netmiko==4.3.0
ntc_templates==4.1.0
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
paramiko==3.4.0
pyarrow==15.0.0
pycparser==2.21
Pygments==2.17.2
PyNaCl==1.5.0
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
scp==0.14.5
six==1.16.0
textfsm==1.1.3
tzdata==2023.4
urllib3==2.2.1
